backend/cloud-functions/behavoural_analysis/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-bigquery==3.*
google-cloud-scheduler>=2.0.0
functions_framework
